requirements.txt
coverage==4.0.3
defusedxml==0.4.1
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.8
django-filter==0.11.0
django-model-utils==2.3.1
django-nose==1.4.3
django-notifications==0.1.dev0
django-notifications-hq==0.8.0
django-secure==1.0.1
django-toolbelt==0.0.1
djangorestframework==3.3.1
djangorestframework-recursive==0.1.1
greenlet==0.4.9
gunicorn==19.4.1
jsonfield==1.0.3
lxml==3.5.0
mock==1.3.0
nose==1.3.7
pbr==1.8.1
proboscis==1.2.6.0
psutil==4.1.0
psycopg2==2.6.1
requests==2.10.0
selenium==2.53.2
six==1.10.0
static3==0.6.1
URLObject==2.4.0
wheel==0.24.0
XlsxWriter==0.8.4
